separate trajectory selector
