+++ planning/autoware_trajectory_selector/launch/trajectory_selector.launch.xml
@@ -10,7 +10,7 @@
   <arg name="input_acceleration" default="~/input/acceleration"/>
   <arg name="input_traffic_signals" default="~/input/traffic_signals"/>
 
-  <node pkg="autoware_trajectory_validator" exec="autoware_trajectory_validator_node" name="trajectory_validator_node" output="screen">
+  <node pkg="autoware_trajectory_selector" exec="autoware_trajectory_selector_node" name="trajectory_selector_node" output="screen">
     <param from="$(var trajectory_validator_param_path)" allow_substs="true"/>
     <remap from="~/input/trajectories_generative" to="$(var input_generative_trajectories)"/>
     <remap from="~/input/trajectories_backup" to="$(var input_backup_trajectories)"/>
