+++ planning/autoware_trajectory_selector/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_trajectory_validator</name>
+  <name>autoware_trajectory_selector</name>
   <version>0.51.0</version>
-  <description>The autoware_trajectory_validator package</description>
+  <description>ROS 2 node that concatenates, validates, and selects candidate trajectories.</description>
 
   <maintainer email="alqudah.mohammad@tier4.jp">Alqudah Mohammad</maintainer>
   <maintainer email="daniel.sanchez@tier4.jp">Daniel Sanchez</maintainer>
@@ -15,39 +15,30 @@
 
   <license>Apache License 2.0</license>
 
-  <author email="go.sakayori@tier4.jp">Go Sakayori</author>
+  <author email="zulfaqar.azmi@tier4.jp">Zulfaqar Azmi</author>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
   <buildtool_depend>autoware_cmake</buildtool_depend>
-  <build_depend>rosidl_default_generators</build_depend>
 
-  <exec_depend>rosidl_default_runtime</exec_depend>
-
-  <depend>autoware_boundary_departure_checker</depend>
   <depend>autoware_internal_planning_msgs</depend>
   <depend>autoware_lanelet2_utils</depend>
   <depend>autoware_map_msgs</depend>
-  <depend>autoware_motion_utils</depend>
   <depend>autoware_perception_msgs</depend>
-  <depend>autoware_planning_msgs</depend>
-  <depend>autoware_traffic_light_utils</depend>
+  <depend>autoware_trajectory_concatenator</depend>
+  <depend>autoware_trajectory_validator</depend>
   <depend>autoware_utils_debug</depend>
-  <depend>autoware_utils_diagnostics</depend>
-  <depend>autoware_utils_geometry</depend>
+  <depend>autoware_utils_rclcpp</depend>
   <depend>autoware_vehicle_info_utils</depend>
-  <depend>builtin_interfaces</depend>
-  <depend>generate_parameter_library</depend>
   <depend>geometry_msgs</depend>
   <depend>nav_msgs</depend>
-  <depend>pluginlib</depend>
   <depend>rclcpp</depend>
+  <depend>rclcpp_components</depend>
   <depend>tl_expected</depend>
-  <depend>unique_identifier_msgs</depend>
 
+  <test_depend>ament_cmake_ros</test_depend>
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>autoware_lint_common</test_depend>
-
-  <member_of_group>rosidl_interface_packages</member_of_group>
+  <test_depend>autoware_test_utils</test_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
